Correct range of day from 01-31 to 00-30
A day of "31" results in the date going to the zeroth day of the next month, even for Feburary, so the range is correct to be from 00 to 30. This has been confirmed on Rush the Rock and Rush 2049SE, it is possible that other games have this error corrected, but unable to test.